test: Cobertura de falha no teste
